Feature for retrieving an active user

+++ src/main/java/com/PBA/authservice/persistance/repository/mappers/PendingUserRowMapperImpl.java
@@ -1,12 +1,10 @@
 package com.pba.authservice.persistance.repository.mappers;
 
 import com.pba.authservice.persistance.model.PendingUser;
-import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Component;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.time.LocalDateTime;
 import java.util.UUID;
 
 @Component
@@ -14,7 +12,7 @@ public class PendingUserRowMapperImpl implements PendingUserRowMapper {
     @Override
     public PendingUser mapRow(ResultSet rs, int rowNum) throws SQLException {
         return PendingUser.builder()
-                .id(rs.getInt("id"))
+                .id(rs.getLong("id"))
                 .uid(UUID.fromString(rs.getString("uid")))
                 .username(rs.getString("username"))
                 .password(rs.getString("password"))
